+++ src/core/sourceType.js
@@ -1,7 +1,12 @@
 /**
  * @param {import('eslint').Rule.RuleContext} context
- * @returns 'module' | 'script' | undefined
+ * @returns 'module' | 'script' | 'commonjs' | undefined
  */
 export default function sourceType(context) {
-  return context.parserOptions.sourceType;
+  if ('sourceType' in context.parserOptions) {
+    return context.parserOptions.sourceType;
+  }
+  if ('languageOptions' in context && context.languageOptions) {
+    return context.languageOptions.sourceType;
+  }
 }